requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
sqlalchemy[asyncio]==2.0.32
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.8.2
python-dotenv==1.0.1
python-multipart